Actualizado README y arreglar tarjetas de remedio
