+++ README.md
@@ -3,50 +3,36 @@
 # rc2-template-proyecto-individual-final
 
 # BOTICA NATURAL
-
 ### App de plantas medicinales y remedios
 
-Dirigida a personas interesadas en la medicina natural y que buscan alivio natural a dolencias comunes y de poca gravedad(dolores decabeza, estres, insomnio, problemas digestivos etc.)
+Una app dirigida a personas interesadas en la medicina natural y que buscan alivio natural a dolencias comunes y de poca gravedad(dolores decabeza, estres, insomnio, problemas digestivos etc.)
 
 # Tecnologías utilizadas
+
+El proyecto utiliza Spring Boot versión 3.2.1 con Maven y Java 17. El acceso a los datos se hace mediante 
+Spring Web, Spring Data JPA y el almacenamiento se realiza en una base de datos H2.
 
 ![Static Badge](https://img.shields.io/badge/React-v%208.2.43-green?logo=react)
 ![JavaScript](https://img.shields.io/badge/JavaScript-ES6-yellow?logo=javascript)
 ![Java](https://img.shields.io/badge/Java-17-blue?logo=java)
 ![Spring Boot](https://img.shields.io/badge/Spring_Boot-3.2.1-green?logo=spring)
 
-# Instalaciones
+# Cómo lanzar el proyecto
 
-##Repository
+## Frontend
+                      
+```
+npm install
+npm run dev 
+```
 
-`git init`
+## Backend
 
-#Frontend
-
-`npm init vite frontend`
-`npm install`
-`npm run dev `
-`npm install --save-dev jest @testing-library/react @testing-library/jest-dom`
-
-#Backend
-
-En barra superior Visual Studio Code
-
-> Spring Initializr:Create Maven Project
-> Spring Version:3.2.1
-> Group Id:
-> Artifact Id:backend
-> Packaging Type:Jar
-> Java Version: 17
-> dependencies:
-> Spring web
-> H2 Database
-> Spring Boot Dev Tools
-> Spring Data JPA
-
-`./mwnw spring-boot:run`
+```
+./mwnw spring-boot:run
+```
 
 # Planificación
 
-- Planificion de tareas en Trello
-- Prototipado en Figma
+- [Planificion de tareas en Trello](enlace)
+- [Prototipado en Figma](enlace)
