+++ backend/pom.xml
@@ -45,7 +45,6 @@
 <dependency>
     <groupId>org.projectlombok</groupId>
     <artifactId>lombok</artifactId>
-    <version>1.18.20</version> <!-- Verifica la última versión en el repositorio de Maven -->
     <scope>provided</scope>
 </dependency>
