+++ frontend/src/App.css
@@ -1,38 +1,13 @@
 :root {
-  --font-size-mobile: 0.8rem;
-  --font-size-tablet: 1rem;
-  --font-size-desktop: 1.1rem;
+    --font-size-mobile: 0.8rem;
+    --font-size-tablet: 1rem;
+    --font-size-desktop: 1.1rem;
 }
 
 body {
-  background: url("../public/assets/fondobotica.jpg") no-repeat center center
-    fixed;
-  background-size: cover;
-  background-color: rgba(255, 255, 255, 1);
-  height: 100vh;
-}
-@media screen and (min-width: 320px) and (max-width: 768px) {
-  .cardRemedio {
-    width: calc(100%-1.25rem);
-    margin-right: 0.625rem;
-    margin-bottom: 0.625rem;
-    font-size: var(--font-size-mobile);
-  }
+    background: url("../public/assets/fondobotica.jpg") no-repeat center center fixed;
+    background-size: cover;
+    background-color: rgba(255, 255, 255, 1);
+    height: 100vh;
 }
 
-@media screen and (min-width: 769px) and (max-width: 1024px) {
-  .cardRemedio {
-    width: calc(50%-1.25rem);
-    margin-right: 0.625rem;
-    margin-bottom: 0.625rem;
-    font-size: var(--font-size-tablet);
-  }
-}
-@media screen and(min-width: 1025px) {
-  .cardRemedio {
-    width: calc(33%-1.25rem);
-    margin-right: 0.625rem;
-    margin-bottom: 0.625rem;
-    font-size: var(--font-size-desktop);
-  }
-}
