--- frontend/src/presentation/components/Cards/CardsRemedios.css
@@ -1,44 +1,0 @@
-.cardRemedio {
-  width: 15%;
-  height: 18.75rem;
-  font-size: 1rem;
-  font-family: "Indie Flower", bold;
-  color: black;
-  margin-right: 2%;
-  margin-bottom: 2%;
-  padding: 0.5rem 1.5rem;
-  border-radius: 0.5rem;
-  background-color: #4f7a4f;
-  float: left;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  transition: transform 0.5s ease-in-out;
-  position: relative;
-  box-shadow: 0.625rem 0.625rem 0.625rem rgb(40, 154, 19);
-  cursor: pointer;
-}
-
-.cardRemedio:hover {
-  transform: scale(1.1);
-}
-.cardRemedioImage {
-  background-repeat: no-repeat;
-  background-size: cover;
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: -1;
-}
-.cardRemedioImage::after {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(10, 10, 10, 0.1);
-  border-radius: 0.5rem;
-}
